Move similar scales template code and styles to their own component
