+++ src/views/AnalysisView.vue
@@ -18,6 +18,7 @@ import ChordWheel from '@/components/ChordWheel.vue'
 import HarmonicEntropyPlot from '@/components/HarmonicEntropyPlot.vue'
 import NumericSlider from '@/components/NumericSlider.vue'
 import ScaleLineInput from '@/components/ScaleLineInput.vue'
+import SimilarScales from '@/components/SimilarScales.vue'
 import { computed, reactive, ref, watch, watchEffect } from 'vue'
 import { useAudioStore } from '@/stores/audio'
 import { useStateStore } from '@/stores/state'
@@ -26,8 +27,6 @@ import { Fraction, lcm, mmod } from 'xen-dev-utils/fraction'
 import { useScaleStore } from '@/stores/scale'
 import { OCTAVE, UNISON } from '@/constants'
 import { useHarmonicEntropyStore } from '@/stores/harmonic-entropy'
-import { useSimilarStore } from '@/stores/similar'
-import { stemBasename } from '@narenratan/scale-library'
 import Values from 'values.js'
 
 const EPSILON = 1e-6
@@ -36,7 +35,6 @@ const audio = useAudioStore()
 const state = useStateStore()
 const scale = useScaleStore()
 const entropy = useHarmonicEntropyStore()
-const similar = useSimilarStore()
 
 const subtab = ref<'matrix' | 'wheels' | 'entropy' | 'similar'>('matrix')
 const cellFormat = ref<'best' | 'fraction' | 'cents' | 'et' | 'decimal'>('best')
@@ -312,13 +310,9 @@ const colors = computed(() =>
   centsValues.value.map((_, i) => scale.colorForIndex(scale.baseMidiNote + entropyMode.value + i))
 )
 
-watch(subtab, async (newValue) => {
+watch(subtab, (newValue) => {
   if (newValue === 'entropy') {
     void entropy.fetchTable()
-  }
-  if (newValue === 'similar') {
-    await similar.fetchLibrary()
-    similar.runComputation(scale.relativeIntervals.map((i) => i.totalCents(true)))
   }
 })
 </script>
@@ -734,102 +728,7 @@ watch(subtab, async (newValue) => {
         </table>
       </div>
     </main>
-    <main v-if="subtab === 'similar'" class="similar-scales">
-      <p v-if="similar.isFetching">Fetching scale library data…</p>
-      <p v-else-if="similar.fetchError" class="error">
-        Failed to load scale library: {{ similar.fetchError }}
-      </p>
-      <template v-else-if="similar.result">
-        <section>
-          <h2>Similar scales</h2>
-          <p v-if="!similar.result.similar.length">No similar scales found.</p>
-          <table v-else>
-            <thead>
-              <tr>
-                <th>File</th>
-                <th>Notes</th>
-                <th>Rotation</th>
-                <th>Max diff (¢)</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr v-for="entry in similar.result.similar" :key="entry.stem">
-                <td>
-                  <a
-                    :href="`https://scalelibrary.org/scales/${entry.stem}/`"
-                    target="_blank"
-                    rel="noopener"
-                    class="external-link"
-                    >{{ stemBasename(entry.stem) }}</a
-                  >
-                </td>
-                <td>{{ entry.notes }}</td>
-                <td>{{ entry.mode }}</td>
-                <td>{{ entry.maxDiff.toFixed(1) }}</td>
-              </tr>
-            </tbody>
-          </table>
-        </section>
-
-        <section>
-          <h2>Child scales</h2>
-          <p v-if="!similar.result.children.length">No child scales found.</p>
-          <table v-else>
-            <thead>
-              <tr>
-                <th>File</th>
-                <th>Notes</th>
-                <th>Max diff (¢)</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr v-for="entry in similar.result.children" :key="entry.stem">
-                <td>
-                  <a
-                    :href="`https://scalelibrary.org/scales/${entry.stem}/`"
-                    target="_blank"
-                    rel="noopener"
-                    class="external-link"
-                    >{{ stemBasename(entry.stem) }}</a
-                  >
-                </td>
-                <td>{{ entry.notes }}</td>
-                <td>{{ entry.maxDiff.toFixed(1) }}</td>
-              </tr>
-            </tbody>
-          </table>
-        </section>
-
-        <section>
-          <h2>Parent scales</h2>
-          <p v-if="!similar.result.parents.length">No parent scales found.</p>
-          <table v-else>
-            <thead>
-              <tr>
-                <th>File</th>
-                <th>Notes</th>
-                <th>Max diff (¢)</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr v-for="entry in similar.result.parents" :key="entry.stem">
-                <td>
-                  <a
-                    :href="`https://scalelibrary.org/scales/${entry.stem}/`"
-                    target="_blank"
-                    rel="noopener"
-                    class="external-link"
-                    >{{ stemBasename(entry.stem) }}</a
-                  >
-                </td>
-                <td>{{ entry.notes }}</td>
-                <td>{{ entry.maxDiff.toFixed(1) }}</td>
-              </tr>
-            </tbody>
-          </table>
-        </section>
-      </template>
-    </main>
+    <SimilarScales v-if="subtab === 'similar'" :relativeIntervals="scale.relativeIntervals" />
   </div>
 </template>
 
@@ -942,38 +841,6 @@ main {
 .entropy-intervals table {
   border-collapse: collapse;
   text-align: center;
-}
-
-/* Similar scales */
-.similar-scales {
-  padding: 0.5rem 1rem;
-  overflow-y: auto;
-}
-.similar-scales section {
-  margin-bottom: 1.5rem;
-}
-.similar-scales table {
-  border-collapse: collapse;
-}
-.similar-scales th,
-.similar-scales td {
-  border: 1px solid var(--color-border);
-  padding: 0.2rem 0.6rem;
-}
-.similar-scales th {
-  font-weight: bold;
-}
-.similar-scales a {
-  color: var(--color-accent-text-btn);
-}
-.similar-scales .error {
-  color: var(--color-error);
-}
-
-a.external-link::after {
-  content: '🔗';
-  display: inline-block;
-  margin-left: 0.25rem;
 }
 
 /* Content layout (medium) */
